feat(Producer/FFXIV)!: Initial commit
