+++ FFXIVDiscordChatBridge/Extensions/DiscordLogSeverityExtensions.cs
@@ -1,6 +1,6 @@
 namespace FFXIVDiscordChatBridge.Extensions;
 
-static class LoggerExtensions
+static class DiscordLogSeverityExtensions
 {
     public static NLog.LogLevel ToNLogSeverity(this Discord.LogSeverity discordLogSeverity)
     {
