+++ FFXIVDiscordChatBridge/NLog.config
@@ -5,10 +5,12 @@
     <targets>
         <target name="logfile" xsi:type="File" fileName="FFXIVDiscordChatBridge.log" />
         <target name="logconsole" xsi:type="Console" />
+        <target name="logdebug" xsi:type="Debugger" />
     </targets>
 
     <rules>
         <logger name="*" minlevel="Trace" writeTo="logconsole" />
         <logger name="*" minlevel="Trace" writeTo="logfile" />
+        <logger name="*" minlevel="Debug" writeTo="logdebug" />
     </rules>
 </nlog>
