+++ FFXIVDiscordChatBridge/Program.cs
@@ -1,25 +1,8 @@
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using NLog;
 
 namespace FFXIVDiscordChatBridge
 {
-    static class ServiceExtensions
-    {
-        public static void AddCommandLineConfiguration(this IServiceCollection services)
-        {
-            var args = Environment.GetCommandLineArgs();
-            var options = args.Where(arg => arg.StartsWith("--"))
-                .Select(arg => arg[2..].Split('='))
-                .Select(arg => new KeyValuePair<string, string?>(arg[0], arg[1]));
-            
-            services.AddSingleton<IConfiguration>(_ => new ConfigurationBuilder()
-                .AddInMemoryCollection(options)
-                .Build());
-            
-            services.AddLogging();
-        }
-    }
     static class Program
     {
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
